http_proxy: move dynhds_add_custom here from http.c
... and make it static. As it is not used anywhere else.
